Split interpolants into a separate file and add comments.
