+++ quartical/interpolation/interpolate.py
@@ -4,9 +4,9 @@ import dask.array as da
 import numpy as np
 import xarray
 import pathlib
-from scipy.interpolate import interp2d
-from csaps import csaps
-from numba import jit
+from quartical.interpolation.interpolants import (interpolate_missing,
+                                                  spline2d_interpolate_gains,
+                                                  csaps2d_interpolate_gains)
 
 
 def load_and_interpolate_gains(gain_xds_list, opts):
@@ -237,7 +237,7 @@ def make_interp_xds_list(term_xds_list, concat_xds_list, interp_mode,
         # This fills in missing values using linear interpolation, or by
         # padding with the last good value (edges). Regions with no good data
         # will be zeroed.
-        interp_xds = interpolate_missing(interp_xds, term_xds, interp_mode)
+        interp_xds = interpolate_missing(interp_xds)
 
         # We may be interpolating from one set of axes to another.
         i_t_axis, i_f_axis = interp_xds.GAIN_AXES[:2]
@@ -251,12 +251,10 @@ def make_interp_xds_list(term_xds_list, concat_xds_list, interp_mode,
                 kwargs={"fill_value": "extrapolate"})
         elif interp_method == "2dspline":
             interp_xds = spline2d_interpolate_gains(interp_xds,
-                                                    term_xds,
-                                                    interp_mode)
+                                                    term_xds)
         elif interp_method == "smoothingspline":
             interp_xds = csaps2d_interpolate_gains(interp_xds,
-                                                   term_xds,
-                                                   interp_mode)
+                                                   term_xds)
 
         # Convert the interpolated quantities back in gains.
         if interp_mode == "ampphase":
@@ -276,192 +274,3 @@ def make_interp_xds_list(term_xds_list, concat_xds_list, interp_mode,
         interp_xds_list.append(interp_xds)
 
     return interp_xds_list
-
-
-def spline2d(x, y, z, xx, yy):
-
-    n_t, n_f, n_a, n_d, n_c = z.shape
-    n_ti, n_fi = xx.size, yy.size
-
-    zz = np.zeros((n_ti, n_fi, n_a, n_d, n_c), dtype=z.dtype)
-
-    # NOTE: x are the column coordinates and y and row coordinates.
-    for a in range(n_a):
-        for d in range(n_d):
-            for c in range(n_c):
-                z_sel = z[:, :, a, d, c]
-                if not np.any(z_sel):
-                    continue
-                interp_func = interp2d(y, x, z_sel, kind="cubic")
-                zz[:, :, a, d, c] = interp_func(yy, xx).reshape(n_ti, n_fi)
-
-    return zz
-
-
-def spline2d_interpolate_gains(interp_xds, term_xds, interp_mode):
-
-    if interp_mode == "ampphase":
-        data_fields = ["amp", "phase"]
-    elif interp_mode == "reim":
-        data_fields = ["re", "im"]
-
-    output_xds = term_xds
-    i_t_axis, i_f_axis = interp_xds.GAIN_AXES[:2]
-    t_t_axis, t_f_axis = output_xds.GAIN_AXES[:2]
-
-    for data_field in data_fields:
-        interp = da.blockwise(spline2d, "tfadc",
-                              interp_xds[i_t_axis].values, None,
-                              interp_xds[i_f_axis].values, None,
-                              interp_xds[data_field].data, "tfadc",
-                              term_xds[t_t_axis].values, None,
-                              term_xds[t_f_axis].values, None,
-                              dtype=np.float64,
-                              adjust_chunks={"t": term_xds.dims[t_t_axis],
-                                             "f": term_xds.dims[t_f_axis]})
-
-        output_xds = output_xds.assign(
-            {data_field: (term_xds.GAIN_AXES, interp)})
-
-    return output_xds
-
-
-def csaps2d(x, y, z, xx, yy):
-
-    n_t, n_f, n_a, n_d, n_c = z.shape
-    n_ti, n_fi = xx.size, yy.size
-
-    zz = np.zeros((n_ti, n_fi, n_a, n_d, n_c), dtype=z.dtype)
-
-    for a in range(n_a):
-        for d in range(n_d):
-            for c in range(n_c):
-                z_sel = z[:, :, a, d, c]
-                if not np.any(z_sel):
-                    continue
-                interp_vals = csaps([x, y], z_sel, [xx, yy]).values
-                zz[:, :, a, d, c] = interp_vals.reshape(n_ti, n_fi)
-
-    return zz
-
-
-def csaps2d_interpolate_gains(interp_xds, term_xds, interp_mode):
-
-    if interp_mode == "ampphase":
-        data_fields = ["amp", "phase"]
-    elif interp_mode == "reim":
-        data_fields = ["re", "im"]
-
-    output_xds = term_xds
-    i_t_axis, i_f_axis = interp_xds.GAIN_AXES[:2]
-    t_t_axis, t_f_axis = output_xds.GAIN_AXES[:2]
-
-    for data_field in data_fields:
-        interp = da.blockwise(csaps2d, "tfadc",
-                              interp_xds[i_t_axis].values, None,
-                              interp_xds[i_f_axis].values, None,
-                              interp_xds[data_field].data, "tfadc",
-                              term_xds[t_t_axis].values, None,
-                              term_xds[t_f_axis].values, None,
-                              dtype=np.float64,
-                              adjust_chunks={"t": term_xds.dims[t_t_axis],
-                                             "f": term_xds.dims[t_f_axis]})
-
-        output_xds = output_xds.assign(
-            {data_field: (term_xds.GAIN_AXES, interp)})
-
-    return output_xds
-
-
-@jit(nopython=True, nogil=True, cache=True)
-def fillmissing(x1, x2, y):
-
-    n_t, n_f, n_a, n_d, n_c = y.shape
-
-    yy = y.copy()
-
-    for f in range(n_f):
-        for a in range(n_a):
-            for d in range(n_d):
-                for c in range(n_c):
-                    y_sel = y[:, f, a, d, c]
-                    good_data = np.where(np.isfinite(y_sel))
-                    if len(good_data[0]) == 0:
-                        continue
-
-                    yy[:, f, a, d, c] = linterp(x1,
-                                                x1[good_data],
-                                                y_sel[good_data])
-
-    for t in range(n_t):
-        for a in range(n_a):
-            for d in range(n_d):
-                for c in range(n_c):
-                    y_sel = yy[t, :, a, d, c]
-                    good_data = np.where(np.isfinite(y_sel))
-                    if len(good_data[0]) == 0:
-                        # If there is no good data along frequency after
-                        # interpolating in time, we have no information
-                        # from which to interpolate - we zero these locations.
-                        yy[:, f, a, d, c] = 0
-                        continue
-
-                    yy[t, :, a, d, c] = linterp(x2,
-                                                x2[good_data],
-                                                y_sel[good_data])
-
-    return yy
-
-
-def interpolate_missing(interp_xds, term_xds, interp_mode):
-
-    if interp_mode == "ampphase":
-        data_fields = ["amp", "phase"]
-    elif interp_mode == "reim":
-        data_fields = ["re", "im"]
-
-    t_axis, f_axis = interp_xds.GAIN_AXES[:2]
-
-    for data_field in data_fields:
-
-        interp = da.blockwise(fillmissing, "tfadc",
-                              interp_xds[t_axis].values, None,
-                              interp_xds[f_axis].values, None,
-                              interp_xds[data_field].data, "tfadc",
-                              dtype=np.float64)
-
-        interp_xds = interp_xds.assign(
-            {data_field: (interp_xds[data_field].dims, interp)})
-
-    return interp_xds
-
-
-@jit(nopython=True, nogil=True, cache=True)
-def linterp(xx, x, y):
-
-    xi = 0
-    xxi = 0
-
-    yy = np.zeros(xx.shape, dtype=y.dtype)
-    xxn = len(xx)
-
-    while xxi < xxn:
-        xxel = xx[xxi]
-        xel = x[xi]
-        if xxel == xel:
-            yy[xxi] = y[xi]
-            xxi += 1
-        elif xxel < x[0]:
-            yy[xxi] = y[0]
-            xxi += 1
-        elif xxel > x[-1]:
-            yy[xxi] = y[-1]
-            xxi += 1
-        elif (xxel > xel) & (xxel < x[xi + 1]):
-            slope = (y[xi + 1] - y[xi]) / (x[xi + 1] - xel)
-            yy[xxi] = slope * (xxel - xel) + y[xi]
-            xxi += 1
-        else:
-            xi += 1
-
-    return yy
